Refactored chat functions into new header files.
